refactor to router and controller
